rename setup->setup_events
