+++ samples/interactive-tutorials/events/import_user_events_gcs_test.py
@@ -15,7 +15,7 @@
 import re
 import subprocess
 
-from setup.setup_cleanup import create_bucket, delete_bucket, upload_blob
+from setup_events.setup_cleanup import create_bucket, delete_bucket, upload_blob
 
 
 def test_import_events_gcs(bucket_name_prefix):
@@ -23,7 +23,7 @@ def test_import_events_gcs(bucket_name_prefix):
     bucket_name = bucket_name_prefix[63:]
 
     try:
-        subprocess.run("python setup/update_user_events_json.py")
+        subprocess.run("python setup_events/update_user_events_json.py")
         create_bucket(bucket_name)
         upload_blob(bucket_name, "../resources/user_events.json")
