+++ samples/interactive-tutorials/events/setup_events/events_create_bigquery_table.py
