+++ samples/interactive-tutorials/events/setup_events/events_delete_gcs_bucket.py
