+++ samples/interactive-tutorials/events/setup_events/setup_cleanup.py
