+++ samples/interactive-tutorials/events/setup_events/update_user_events_json.py
